kbacklight: Add check for BRIGHTNESS_FILE

+++ apps/kbacklight.bash
@@ -12,6 +12,13 @@ if [[ $# != 1 ]]; then
     exit 1
 fi
 
+if [[ ! -f "$BRIGHTNESS_FILE" ]]; then
+    echo "Brightness file doesn't exist! ($BRIGHTNESS_FILE)" >&2
+    echo "Reload the acpi module." >&2
+    notify-send "Kbacklight" "Brightness file doesn't exist! Reload the acpi module."
+    exit 1
+fi
+
 case "$1" in
     0|1|2)
         echo -n "$1" > "$BRIGHTNESS_FILE"
